Build the logical tag tree in a pre-pass
The whole logical tree structure is then available in a second pass
that populates the tree with content.

This is necessary because of a couple of requirements:
- The frame tree must be processed page by page, and in order
  because the relative transforms of each nested group frame
  must be chained.
- When writing marked content sequences, we must immediately decide
  whether the content is an artifacts or not.
If a frame that has a logical parent set, would turn out to be inside
an artifact, we would have to know that (possibly) before processing
the parent frame.